Refactored AbstractDBCPConnectionPool, moved class specific code to subclasses.
