+++ nifi-nar-bundles/nifi-extension-utils/nifi-dbcp-base/src/main/java/org/apache/nifi/dbcp/AbstractDBCPConnectionPool.java
@@ -23,258 +23,37 @@ import org.apache.nifi.annotation.lifecycle.OnEnabled;
 import org.apache.nifi.components.ConfigVerificationResult;
 import org.apache.nifi.components.PropertyDescriptor;
 import org.apache.nifi.components.PropertyValue;
-import org.apache.nifi.components.resource.ResourceCardinality;
-import org.apache.nifi.components.resource.ResourceType;
 import org.apache.nifi.controller.AbstractControllerService;
 import org.apache.nifi.controller.ConfigurationContext;
 import org.apache.nifi.controller.VerifiableControllerService;
-import org.apache.nifi.expression.AttributeExpression;
-import org.apache.nifi.expression.ExpressionLanguageScope;
-import org.apache.nifi.kerberos.KerberosCredentialsService;
+import org.apache.nifi.dbcp.utils.DBCPProperties;
+import org.apache.nifi.dbcp.utils.DataSourceConfiguration;
 import org.apache.nifi.kerberos.KerberosUserService;
 import org.apache.nifi.logging.ComponentLog;
 import org.apache.nifi.processor.exception.ProcessException;
-import org.apache.nifi.processor.util.StandardValidators;
 import org.apache.nifi.reporting.InitializationException;
 import org.apache.nifi.security.krb.KerberosAction;
-import org.apache.nifi.security.krb.KerberosKeytabUser;
 import org.apache.nifi.security.krb.KerberosLoginException;
-import org.apache.nifi.security.krb.KerberosPasswordUser;
 import org.apache.nifi.security.krb.KerberosUser;
 
 import java.sql.Connection;
 import java.sql.Driver;
-import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
 import static org.apache.nifi.components.ConfigVerificationResult.Outcome.FAILED;
 import static org.apache.nifi.components.ConfigVerificationResult.Outcome.SUCCESSFUL;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_EVICTION_RUN_PERIOD;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MAX_CONN_LIFETIME;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MAX_IDLE;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MAX_TOTAL_CONNECTIONS;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MAX_WAIT_MILLIS;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MIN_EVICTABLE_IDLE_TIME_MINS;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_MIN_IDLE;
-import static org.apache.nifi.dbcp.BasicDataSourceConfiguration.DEFAULT_SOFT_MIN_EVICTABLE_IDLE_TIME;
+import static org.apache.nifi.dbcp.utils.DBCPProperties.DB_DRIVERNAME;
+import static org.apache.nifi.dbcp.utils.DBCPProperties.DB_DRIVER_LOCATION;
+import static org.apache.nifi.dbcp.utils.DBCPProperties.KERBEROS_USER_SERVICE;
 
-/**
- * Abstract base class for Database Connection Pooling Services using Apache Commons DBCP as the underlying connection pool implementation.
- */
 public abstract class AbstractDBCPConnectionPool extends AbstractControllerService implements DBCPService, VerifiableControllerService {
-    protected static final String SENSITIVE_PROPERTY_PREFIX = "SENSITIVE.";
-    public static final PropertyDescriptor DATABASE_URL = new PropertyDescriptor.Builder()
-            .name("Database Connection URL")
-            .description("A database connection URL used to connect to a database. May contain database system name, host, port, database name and some parameters."
-                    + " The exact syntax of a database connection URL is specified by your DBMS.")
-            .defaultValue(null)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .required(true)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor DB_DRIVERNAME = new PropertyDescriptor.Builder()
-            .name("Database Driver Class Name")
-            .description("Database driver class name")
-            .defaultValue(null)
-            .required(true)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor DB_DRIVER_LOCATION = new PropertyDescriptor.Builder()
-            .name("database-driver-locations")
-            .displayName("Database Driver Location(s)")
-            .description("Comma-separated list of files/folders and/or URLs containing the driver JAR and its dependencies (if any). For example '/var/tmp/mariadb-java-client-1.1.7.jar'")
-            .defaultValue(null)
-            .required(false)
-            .identifiesExternalResource(ResourceCardinality.MULTIPLE, ResourceType.FILE, ResourceType.DIRECTORY, ResourceType.URL)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .dynamicallyModifiesClasspath(true)
-            .build();
-
-    public static final PropertyDescriptor DB_USER = new PropertyDescriptor.Builder()
-            .name("Database User")
-            .description("Database user name")
-            .defaultValue(null)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor DB_PASSWORD = new PropertyDescriptor.Builder()
-            .name("Password")
-            .description("The password for the database user")
-            .defaultValue(null)
-            .required(false)
-            .sensitive(true)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor MAX_WAIT_TIME = new PropertyDescriptor.Builder()
-            .name("Max Wait Time")
-            .description("The maximum amount of time that the pool will wait (when there are no available connections) "
-                    + " for a connection to be returned before failing, or -1 to wait indefinitely. ")
-            .defaultValue(DEFAULT_MAX_WAIT_MILLIS)
-            .required(true)
-            .addValidator(DBCPValidator.CUSTOM_TIME_PERIOD_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .sensitive(false)
-            .build();
-
-    public static final PropertyDescriptor MAX_TOTAL_CONNECTIONS = new PropertyDescriptor.Builder()
-            .name("Max Total Connections")
-            .description("The maximum number of active connections that can be allocated from this pool at the same time, "
-                    + " or negative for no limit.")
-            .defaultValue(DEFAULT_MAX_TOTAL_CONNECTIONS)
-            .required(true)
-            .addValidator(StandardValidators.INTEGER_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .sensitive(false)
-            .build();
-
-    public static final PropertyDescriptor VALIDATION_QUERY = new PropertyDescriptor.Builder()
-            .name("Validation-query")
-            .displayName("Validation query")
-            .description("Validation query used to validate connections before returning them. "
-                    + "When connection is invalid, it gets dropped and new valid connection will be returned. "
-                    + "Note!! Using validation might have some performance penalty.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor MIN_IDLE = new PropertyDescriptor.Builder()
-            .displayName("Minimum Idle Connections")
-            .name("dbcp-min-idle-conns")
-            .description("The minimum number of connections that can remain idle in the pool without extra ones being " +
-                    "created. Set to or zero to allow no idle connections.")
-            .defaultValue(DEFAULT_MIN_IDLE)
-            .required(false)
-            .addValidator(StandardValidators.NON_NEGATIVE_INTEGER_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor MAX_IDLE = new PropertyDescriptor.Builder()
-            .displayName("Max Idle Connections")
-            .name("dbcp-max-idle-conns")
-            .description("The maximum number of connections that can remain idle in the pool without extra ones being " +
-                    "released. Set to any negative value to allow unlimited idle connections.")
-            .defaultValue(DEFAULT_MAX_IDLE)
-            .required(false)
-            .addValidator(StandardValidators.INTEGER_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor MAX_CONN_LIFETIME = new PropertyDescriptor.Builder()
-            .displayName("Max Connection Lifetime")
-            .name("dbcp-max-conn-lifetime")
-            .description("The maximum lifetime in milliseconds of a connection. After this time is exceeded the " +
-                    "connection will fail the next activation, passivation or validation test. A value of zero or less " +
-                    "means the connection has an infinite lifetime.")
-            .defaultValue(DEFAULT_MAX_CONN_LIFETIME)
-            .required(false)
-            .addValidator(DBCPValidator.CUSTOM_TIME_PERIOD_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor EVICTION_RUN_PERIOD = new PropertyDescriptor.Builder()
-            .displayName("Time Between Eviction Runs")
-            .name("dbcp-time-between-eviction-runs")
-            .description("The number of milliseconds to sleep between runs of the idle connection evictor thread. When " +
-                    "non-positive, no idle connection evictor thread will be run.")
-            .defaultValue(DEFAULT_EVICTION_RUN_PERIOD)
-            .required(false)
-            .addValidator(DBCPValidator.CUSTOM_TIME_PERIOD_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor MIN_EVICTABLE_IDLE_TIME = new PropertyDescriptor.Builder()
-            .displayName("Minimum Evictable Idle Time")
-            .name("dbcp-min-evictable-idle-time")
-            .description("The minimum amount of time a connection may sit idle in the pool before it is eligible for eviction.")
-            .defaultValue(DEFAULT_MIN_EVICTABLE_IDLE_TIME_MINS)
-            .required(false)
-            .addValidator(DBCPValidator.CUSTOM_TIME_PERIOD_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor SOFT_MIN_EVICTABLE_IDLE_TIME = new PropertyDescriptor.Builder()
-            .displayName("Soft Minimum Evictable Idle Time")
-            .name("dbcp-soft-min-evictable-idle-time")
-            .description("The minimum amount of time a connection may sit idle in the pool before it is eligible for " +
-                    "eviction by the idle connection evictor, with the extra condition that at least a minimum number of" +
-                    " idle connections remain in the pool. When the not-soft version of this option is set to a positive" +
-                    " value, it is examined first by the idle connection evictor: when idle connections are visited by " +
-                    "the evictor, idle time is first compared against it (without considering the number of idle " +
-                    "connections in the pool) and then against this soft option, including the minimum idle connections " +
-                    "constraint.")
-            .defaultValue(DEFAULT_SOFT_MIN_EVICTABLE_IDLE_TIME)
-            .required(false)
-            .addValidator(DBCPValidator.CUSTOM_TIME_PERIOD_VALIDATOR)
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor KERBEROS_CREDENTIALS_SERVICE = new PropertyDescriptor.Builder()
-            .name("kerberos-credentials-service")
-            .displayName("Kerberos Credentials Service")
-            .description("Specifies the Kerberos Credentials Controller Service that should be used for authenticating with Kerberos")
-            .identifiesControllerService(KerberosCredentialsService.class)
-            .required(false)
-            .build();
-
-    public static final PropertyDescriptor KERBEROS_USER_SERVICE = new PropertyDescriptor.Builder()
-            .name("kerberos-user-service")
-            .displayName("Kerberos User Service")
-            .description("Specifies the Kerberos User Controller Service that should be used for authenticating with Kerberos")
-            .identifiesControllerService(KerberosUserService.class)
-            .required(false)
-            .build();
-
-    public static final PropertyDescriptor KERBEROS_PRINCIPAL = new PropertyDescriptor.Builder()
-            .name("kerberos-principal")
-            .displayName("Kerberos Principal")
-            .description("The principal to use when specifying the principal and password directly in the processor for authenticating via Kerberos.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .addValidator(StandardValidators.createAttributeExpressionLanguageValidator(AttributeExpression.ResultType.STRING))
-            .expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY)
-            .build();
-
-    public static final PropertyDescriptor KERBEROS_PASSWORD = new PropertyDescriptor.Builder()
-            .name("kerberos-password")
-            .displayName("Kerberos Password")
-            .description("The password to use when specifying the principal and password directly in the processor for authenticating via Kerberos.")
-            .required(false)
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .sensitive(true)
-            .build();
 
     protected volatile BasicDataSource dataSource;
     protected volatile KerberosUser kerberosUser;
-
-    @Override
-    protected PropertyDescriptor getSupportedDynamicPropertyDescriptor(final String propertyDescriptorName) {
-        final PropertyDescriptor.Builder builder = new PropertyDescriptor.Builder()
-                .name(propertyDescriptorName)
-                .required(false)
-                .dynamic(true)
-                .addValidator(StandardValidators.createAttributeExpressionLanguageValidator(AttributeExpression.ResultType.STRING, true))
-                .addValidator(StandardValidators.ATTRIBUTE_KEY_PROPERTY_NAME_VALIDATOR);
-
-        if (propertyDescriptorName.startsWith(SENSITIVE_PROPERTY_PREFIX)) {
-            builder.sensitive(true).expressionLanguageSupported(ExpressionLanguageScope.NONE);
-        } else {
-            builder.expressionLanguageSupported(ExpressionLanguageScope.VARIABLE_REGISTRY);
-        }
-
-        return builder.build();
-    }
 
     @Override
     public List<ConfigVerificationResult> verify(final ConfigurationContext context, final ComponentLog verificationLogger, final Map<String, String> variables) {
@@ -301,7 +80,8 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
 
         final BasicDataSource dataSource = new BasicDataSource();
         try {
-            configureDataSource(kerberosUser, context);
+            final DataSourceConfiguration configuration = createDataSourceConfiguraton(context);
+            configureDataSource(context, configuration);
             results.add(new ConfigVerificationResult.Builder()
                     .verificationStepName("Configure Data Source")
                     .outcome(SUCCESSFUL)
@@ -341,7 +121,6 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
                 verificationLogger.error("Failed to shut down data source", e);
             }
         }
-
         return results;
     }
 
@@ -359,13 +138,14 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
      */
     @OnEnabled
     public void onConfigured(final ConfigurationContext context) throws InitializationException {
-        kerberosUser = getKerberosUser(context);
         dataSource = new BasicDataSource();
-        configureDataSource(kerberosUser, context);
+        kerberosUser = getKerberosUser(context);
+        loginKerberos(kerberosUser);
+        final DataSourceConfiguration configuration = createDataSourceConfiguraton(context);
+        configureDataSource(context, configuration);
     }
 
-    protected void configureDataSource(final KerberosUser kerberosUser, final ConfigurationContext context) throws InitializationException {
-
+    private void loginKerberos(KerberosUser kerberosUser) throws InitializationException {
         if (kerberosUser != null) {
             try {
                 kerberosUser.login();
@@ -373,48 +153,24 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
                 throw new InitializationException("Unable to authenticate Kerberos principal", e);
             }
         }
-
-        final String dbUrl = getUrl(context);
-        final String driverName = context.getProperty(DB_DRIVERNAME).evaluateAttributeExpressions().getValue();
-        final Driver driver = getDriver(driverName, dbUrl);
-
-        final String userName = context.getProperty(DB_USER).evaluateAttributeExpressions().getValue();
-        final String password = context.getProperty(DB_PASSWORD).evaluateAttributeExpressions().getValue();
-        final String validationQuery = context.getProperty(VALIDATION_QUERY).evaluateAttributeExpressions().getValue();
-
-        final PropertyValue maxTotal = context.getProperty(MAX_TOTAL_CONNECTIONS);
-        final PropertyValue maxWaitMillis = context.getProperty(MAX_WAIT_TIME);
-        final PropertyValue minIdle = context.getProperty(MIN_IDLE);
-        final PropertyValue maxIdle = context.getProperty(MAX_IDLE);
-        final PropertyValue maxConnLifetimeMillis = context.getProperty(MAX_CONN_LIFETIME);
-        final PropertyValue timeBetweenEvictionRunsMillis = context.getProperty(EVICTION_RUN_PERIOD);
-        final PropertyValue minEvictableIdleTimeMillis = context.getProperty(MIN_EVICTABLE_IDLE_TIME);
-        final PropertyValue softMinEvictableIdleTimeMillis = context.getProperty(SOFT_MIN_EVICTABLE_IDLE_TIME);
-
-        final List<PropertyDescriptor> dynamicProperties = context.getProperties()
-                .keySet()
-                .stream()
-                .filter(PropertyDescriptor::isDynamic)
-                .collect(Collectors.toList());
-
-        final BasicDataSourceConfiguration basicDataSourceConfiguration = new BasicDataSourceConfiguration.Builder(dbUrl, driver, userName, password)
-                .maxTotal(maxTotal)
-                .maxWaitMillis(maxWaitMillis)
-                .minIdle(minIdle)
-                .maxIdle(maxIdle)
-                .maxConnLifetimeMillis(maxConnLifetimeMillis)
-                .timeBetweenEvictionRunsMillis(timeBetweenEvictionRunsMillis)
-                .minEvictableIdleTimeMillis(minEvictableIdleTimeMillis)
-                .softMinEvictableIdleTimeMillis(softMinEvictableIdleTimeMillis)
-                .validationQuery(validationQuery)
-                .dynamicProperties(dynamicProperties)
-                .build();
-
-        configureDataSourceFromConfiguration(context, dataSource, basicDataSourceConfiguration);
     }
 
-    private void configureDataSourceFromConfiguration(final ConfigurationContext context, final BasicDataSource dataSource, final BasicDataSourceConfiguration configuration) {
-        dataSource.setDriver(configuration.getDriver());
+    protected abstract Driver getDriver(final String driverName, final String url);
+
+    protected DataSourceConfiguration createDataSourceConfiguraton(final ConfigurationContext context) {
+        final String url = context.getProperty(DBCPProperties.DATABASE_URL).getValue();
+        final String user = context.getProperty(DBCPProperties.DB_USER).getValue();
+        final String password = context.getProperty(DBCPProperties.DB_PASSWORD).getValue();
+        final String driverName = context.getProperty(DB_DRIVERNAME).evaluateAttributeExpressions().getValue();
+
+        return new DataSourceConfiguration.Builder(url, driverName, user, password)
+                .build();
+    }
+
+    protected void configureDataSource(final ConfigurationContext context, final DataSourceConfiguration configuration) {
+        final Driver driver = getDriver(configuration.getDriverName(), configuration.getUrl());
+
+        dataSource.setDriver(driver);
         dataSource.setMaxWaitMillis(configuration.getMaxWaitMillis());
         dataSource.setMaxTotal(configuration.getMaxTotal());
         dataSource.setMinIdle(configuration.getMinIdle());
@@ -434,101 +190,42 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
         dataSource.setUsername(configuration.getUserName());
         dataSource.setPassword(configuration.getPassword());
 
-        configuration.getDynamicProperties().forEach(descriptor -> {
-            final PropertyValue propertyValue = context.getProperty(descriptor);
-            if (descriptor.isSensitive()) {
-                final String propertyName = StringUtils.substringAfter(descriptor.getName(), SENSITIVE_PROPERTY_PREFIX);
-                dataSource.addConnectionProperty(propertyName, propertyValue.getValue());
-            } else {
-                dataSource.addConnectionProperty(descriptor.getName(), propertyValue.evaluateAttributeExpressions().getValue());
-            }
-        });
-
         getConnectionProperties(context).forEach(dataSource::addConnectionProperty);
     }
 
+    protected Map<String, String> getConnectionProperties(final ConfigurationContext context) {
+        return getDynamicProperties(context)
+                .stream()
+                .collect(Collectors.toMap(PropertyDescriptor::getName, s -> {
+                    final PropertyValue propertyValue = context.getProperty(s);
+                    return propertyValue.evaluateAttributeExpressions().getValue();
+                }));
+    }
+
+    protected List<PropertyDescriptor> getDynamicProperties(final ConfigurationContext context) {
+        return context.getProperties()
+                .keySet()
+                .stream()
+                .filter(PropertyDescriptor::isDynamic)
+                .collect(Collectors.toList());
+    }
+
     protected KerberosUser getKerberosUser(final ConfigurationContext context) {
-        KerberosUser kerberosUser = null;
-        final KerberosCredentialsService kerberosCredentialsService = context.getProperty(KERBEROS_CREDENTIALS_SERVICE).asControllerService(KerberosCredentialsService.class);
+        KerberosUser kerberosUser;
         final KerberosUserService kerberosUserService = context.getProperty(KERBEROS_USER_SERVICE).asControllerService(KerberosUserService.class);
-        final String kerberosPrincipal = context.getProperty(KERBEROS_PRINCIPAL).evaluateAttributeExpressions().getValue();
-        final String kerberosPassword = context.getProperty(KERBEROS_PASSWORD).getValue();
 
         if (kerberosUserService != null) {
             kerberosUser = kerberosUserService.createKerberosUser();
-        } else if (kerberosCredentialsService != null) {
-            kerberosUser = new KerberosKeytabUser(kerberosCredentialsService.getPrincipal(), kerberosCredentialsService.getKeytab());
-        } else if (!StringUtils.isBlank(kerberosPrincipal) && !StringUtils.isBlank(kerberosPassword)) {
-            kerberosUser = new KerberosPasswordUser(kerberosPrincipal, kerberosPassword);
+        } else {
+            kerberosUser = getKerberosUserByCredentials(context);
         }
         return kerberosUser;
     }
 
-    protected String getUrl(ConfigurationContext context) {
-        return context.getProperty(DATABASE_URL).evaluateAttributeExpressions().getValue();
+    protected KerberosUser getKerberosUserByCredentials(final ConfigurationContext context) {
+        return null;
     }
 
-    protected Driver getDriver(final String driverName, final String url) {
-        final Class<?> clazz;
-
-        try {
-            clazz = Class.forName(driverName);
-        } catch (final ClassNotFoundException e) {
-            throw new ProcessException("Driver class " + driverName + " is not found", e);
-        }
-
-        try {
-            return DriverManager.getDriver(url);
-        } catch (final SQLException e) {
-            // In case the driver is not registered by the implementation, we explicitly try to register it.
-            try {
-                final Driver driver = (Driver) clazz.newInstance();
-                DriverManager.registerDriver(driver);
-                return DriverManager.getDriver(url);
-            } catch (final SQLException e2) {
-                throw new ProcessException("No suitable driver for the given Database Connection URL", e2);
-            } catch (final IllegalAccessException | InstantiationException e2) {
-                throw new ProcessException("Creating driver instance is failed", e2);
-            }
-        }
-    }
-
-    /**
-     * Override in subclasses to provide connection properties to the data source
-     *
-     * @return Key-value pairs that will be added as connection properties
-     */
-    protected Map<String, String> getConnectionProperties(final ConfigurationContext context) {
-        return new HashMap<>();
-    }
-
-    /**
-     * Shutdown pool, close all open connections.
-     * If a principal is authenticated with a KDC, that principal is logged out.
-     *
-     * @throws SQLException if there is an error while closing open connections
-     */
-    @OnDisabled
-    public void shutdown() throws SQLException {
-        try {
-            this.shutdown(dataSource, kerberosUser);
-        } finally {
-            kerberosUser = null;
-            dataSource = null;
-        }
-    }
-
-    private void shutdown(final BasicDataSource dataSource, final KerberosUser kerberosUser) throws SQLException {
-        try {
-            if (kerberosUser != null) {
-                kerberosUser.logout();
-            }
-        } finally {
-            if (dataSource != null) {
-                dataSource.close();
-            }
-        }
-    }
 
     @Override
     public Connection getConnection() throws ProcessException {
@@ -559,8 +256,33 @@ public abstract class AbstractDBCPConnectionPool extends AbstractControllerServi
         }
     }
 
-    @Override
-    public String toString() {
-        return this.getClass().getSimpleName() + "[id=" + getIdentifier() + "]";
+    /**
+     * Shutdown pool, close all open connections.
+     * If a principal is authenticated with a KDC, that principal is logged out.
+     *
+     * @throws SQLException if there is an error while closing open connections
+     */
+    @OnDisabled
+    public void shutdown() throws SQLException {
+        try {
+            shutdown(dataSource, kerberosUser);
+        } finally {
+            kerberosUser = null;
+            dataSource = null;
+        }
     }
+
+    private void shutdown(final BasicDataSource dataSource, final KerberosUser kerberosUser) throws SQLException {
+        try {
+            if (kerberosUser != null) {
+                kerberosUser.logout();
+            }
+        } finally {
+            if (dataSource != null) {
+                dataSource.close();
+            }
+        }
+    }
+
+
 }
